+++ src/brom_drake/productions/debug/__init__.py
@@ -1,0 +1,7 @@
+from brom_drake.productions.debug.show_me import (
+    ShowMeThisModel,
+)
+
+__all__ = [
+    'ShowMeThisModel',
+]
