+++ src/brom_drake/productions/motion_planning/offline/__init__.py
@@ -1,0 +1,7 @@
+from .chem_lab1 import ChemLab1
+from .shelf import ShelfPlanning1
+
+__all__ = [
+    "ShelfPlanning1",
+    "ChemLab1",
+]
