+++ src/brom_drake/productions/roles/__init__.py
@@ -1,0 +1,8 @@
+from brom_drake.productions.roles.role import Role, RolePortAssignment
+from brom_drake.productions.roles.motion_planners.kinematic import kKinematicMotionPlanner
+
+__all__ = [
+    "Role",
+    "kKinematicMotionPlanner",
+    "RolePortAssignment",
+]
