+++ src/brom_drake/productions/types/motion_planning/__init__.py
@@ -1,0 +1,3 @@
+from .kinematic import KinematicMotionPlanningProduction
+
+__all__ = ["KinematicMotionPlanningProduction"]
