--- src/brom_drake/scenes/debug/__init__.py
@@ -1,7 +1,0 @@
-from brom_drake.scenes.debug.show_me import (
-    ShowMeThisModel,
-)
-
-__all__ = [
-    'ShowMeThisModel',
-]
