--- src/brom_drake/scenes/motion_planning/offline/__init__.py
@@ -1,7 +1,0 @@
-from .chem_lab1 import ChemLab1Scene
-from .shelf import ShelfPlanningScene
-
-__all__ = [
-    "ShelfPlanningScene",
-    "ChemLab1Scene",
-]
