--- src/brom_drake/scenes/roles/__init__.py
@@ -1,8 +1,0 @@
-from brom_drake.scenes.roles.role import Role, RolePortAssignment
-from brom_drake.scenes.roles.motion_planners.kinematic import kKinematicMotionPlanner
-
-__all__ = [
-    "Role",
-    "kKinematicMotionPlanner",
-    "RolePortAssignment",
-]
