--- src/brom_drake/scenes/types/__init__.py
@@ -1,8 +1,0 @@
-from .base import BaseScene, Performer
-from .motion_planning import KinematicMotionPlanningScene
-
-__all__ = [
-    "BaseScene",
-    "KinematicMotionPlanningScene",
-    "Performer",
-]
