--- src/brom_drake/scenes/types/motion_planning/__init__.py
@@ -1,3 +1,0 @@
-from .kinematic import KinematicMotionPlanningScene
-
-__all__ = ["KinematicMotionPlanningScene"]
